+++ CHANGELOG.md
@@ -4,7 +4,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+- response messages now support markdown by default *(2021-07-21)*
+- added dev mode to print regrex matching in BotResponse.js *(2021-07-21)*
 
 ## [1.0.3]
 ### Updated